% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_survey.R
\name{create_survey}
\alias{create_survey}
\title{create_survey}
\usage{
create_survey(question.type = "Matrix", answer.type = "SingleAnswer",
  orientation = "Vertical", survey.content = list.test,
  matrix.answers = "None", n.questions.per.block = 10,
  filename = "choiceblocks")
}
\arguments{
\item{question.type}{One of 'MC' (default), 'Matrix', 'TextEntry' (or 'TE'), 'ConstantSum' (or 'CS'), 'RankOrder' (or 'RO'), or 'Text' (or 'DB').}

\item{answer.type}{One of 'SingleAnswer' (default), 'MultipleAnswer', 'DropDown', 'Select', 'MultiSelect', 'SingleLine', 'Essay', or 'Form'.}

\item{orientation}{Either 'Vertical' (default) or 'Horizontal' for multiple choice question orientation.}

\item{survey.content}{A list of question choices. The names of the list should be the question text. Choice encoding may be passed by naming the choices.}

\item{matrix.answers}{A vector of answer choices for the matrix question. Answeer encoding can be passed by naming the vector. Defaults to ``None''.}

\item{n.questions.per.block}{The number of questions on each Qualtrics block (or page). Defaults to 10.}

\item{filename}{The name of the output file.}
}
\description{
This function creates Qualtrics surveys ready for import. Users must provide survey questions and choices as a named list where the list names are the questions and the content of each list are the choices. Encoding can be passed by naming the list choices. For matrix questions, users must provide a vector of answers, again passing encoding via naming the vector. For a detailed overview of possible question, choice, and answer types, please see \code{\link{https://www.qualtrics.com/support/survey-platform/survey-module/survey-tools/import-and-export-surveys/}}.
}
\examples{
data("samplesurvey")
create_survey(question.type = "MC",
              answer.type = "SingleAnswer",
              orientation = "Vertical",
              survey.content = samplesurvey,
              matrix.answers = "None",
              n.questions.per.block = 10,
              filename = "MCSurvey.txt")
              
create_survey(question.type = "Matrix",
              answer.type = "MultipleAnswer",
              orientation = "Vertical",
              survey.content = samplesurvey,
              matrix.answers = matrixAnswers,
              n.questions.per.block = 5,
              filename = "MatrixSurvey.txt")
}
\keyword{Qualtrics}
\keyword{automation}
\keyword{creator}
\keyword{survey}
